Update README and handle signals in main thread only
The README.md is updated to include essential information on using 'wingrab' in sub threads. This was done to provide clearer usage instructions for the users.

In 'wingrab.py', signal handling has been moved inside a try-except block to prohibit usage of 'signal' outside of main thread. This ensures that 'wingrab' is imported in the main thread initially.

The use of `_debug` as a keyword-only argument in `grab()` and `cleanup()` functions has been implemented to prevent users from accidentally passing in a positional argument.